+++ lib/lua/visitors/runtime/semantics.dart
@@ -2,4 +2,4 @@ import 'package:puredartlua/lua/visitors/runtime/base.dart';
 
 abstract class Semantics {}
 
-typedef IncludeCallback = Object? Function(String, BaseRuntime);
+typedef LuaRequireCallback = Object? Function(String, BaseRuntime);